Fixed type mismatch issue

+++ brendancuda_ai_mlp_fixedmlp.cuh
@@ -119,7 +119,7 @@ __host__ __device__ void BrendanCUDA::AI::MLP::FixedMLPL<_T, _ActivationFunction
 template <typename _T, BrendanCUDA::AI::activationFunction_t<_T> _ActivationFunction, size_t _InputCount, size_t _OutputCount>
 __host__ __device__ void BrendanCUDA::AI::MLP::FixedMLPL<_T, _ActivationFunction, _InputCount, _OutputCount>::Run(const _T* Input, _T* Output) const {
     for (size_t j = 0; j < _OutputCount; ++j) {
-        size_t v = bias[j];
+        float v = bias[j];
         for (size_t i = 0; i < _InputCount; ++i) {
             v += weights[i][j] * Input[i];
         }
